pyproject.toml
[project]
name = "modularized-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastlite>=0.1.2",
    "requests>=2.32.3",
    "sqlite-vec>=0.1.6",
    "typer>=0.15.1",
    "ipykernel>=6.29.5",
    "ruff>=0.9.9",
    "tokenizers>=0.21.0",
    "tqdm>=4.67.1",
    "matplotlib>=3.10.1",
    "tabulate>=0.9.0",
    "google>=3.0.0",
    "google-generativeai>=0.4.0",
    "ipywidgets>=8.1.5",
    "pandas>=2.2.3",
    "anywidget>=0.9.16",
    "marker-pdf==1.6.1",
    "duckdb>=1.2.1",
    "einops>=0.8.1",
    "seaborn>=0.13.2",
    "scikit-learn>=1.6.1",
    "openai>=1.70.0",
    "transformers>=4.50.3",
    "torch>=2.6.0",
    "sentence-transformers>=4.0.1",
    "python-dotenv>=1.1.0",
]

[tool.uv.workspace]
members = ["myenv"]
